@@ -1,7 +1,6 @@
 import 'dart:async';
 import 'dart:io';
 
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
@@ -11,8 +10,8 @@ class FlutterAppCenter {
 
   /// Start appcenter functionalities
   static Future startAsync({
-    @required String appSecretAndroid,
-    @required String appSecretIOS,
+    required String appSecretAndroid,
+    required String appSecretIOS,
     enableAnalytics = true,
     enableCrashes = true,
     enableDistribute = false,
@@ -28,7 +27,7 @@ class FlutterAppCenter {
       throw UnsupportedError('Current platform is not supported.');
     }
 
-    if (appsecret == null || appsecret.isEmpty) {
+    if (appsecret.isEmpty) {
       return;
     }
 
@@ -50,7 +49,7 @@ class FlutterAppCenter {
 
   /// Track events
   static Future trackEventAsync(String name,
-      [Map<String, String> properties]) async {
+      [Map<String, String>? properties]) async {
     await _methodChannel.invokeMethod('trackEvent', <String, dynamic>{
       'name': name,
       'properties': properties ?? <String, String>{},
@@ -58,7 +57,7 @@ class FlutterAppCenter {
   }
 
   /// Check whether analytics is enalbed
-  static Future<bool> isAnalyticsEnabledAsync() async {
+  static Future<bool?> isAnalyticsEnabledAsync() async {
     return await _methodChannel.invokeMethod('isAnalyticsEnabled');
   }
 
@@ -70,32 +69,32 @@ class FlutterAppCenter {
   }
 
   /// Enable or disable analytics
-  static Future configureAnalyticsAsync({@required enabled}) async {
+  static Future configureAnalyticsAsync({required enabled}) async {
     await _methodChannel.invokeMethod('configureAnalytics', enabled);
   }
 
   /// Check whether crashes is enabled
-  static Future<bool> isCrashesEnabledAsync() async {
+  static Future<bool?> isCrashesEnabledAsync() async {
     return await _methodChannel.invokeMethod('isCrashesEnabled');
   }
 
   /// Enable or disable appcenter crash reports
-  static Future configureCrashesAsync({@required enabled}) async {
+  static Future configureCrashesAsync({required enabled}) async {
     await _methodChannel.invokeMethod('configureCrashes', enabled);
   }
 
   /// Check whether appcenter distribution is enabled
-  static Future<bool> isDistributeEnabledAsync() async {
+  static Future<bool?> isDistributeEnabledAsync() async {
     return await _methodChannel.invokeMethod('isDistributeEnabled');
   }
 
   /// Enable or disable appcenter distribution
-  static Future configureDistributeAsync({@required enabled}) async {
+  static Future configureDistributeAsync({required enabled}) async {
     await _methodChannel.invokeMethod('configureDistribute', enabled);
   }
 
   /// Enable or disable appcenter distribution for debug build (Android only)
-  static Future configureDistributeDebugAsync({@required enabled}) async {
+  static Future configureDistributeDebugAsync({required enabled}) async {
     await _methodChannel.invokeMethod('configureDistributeDebug', enabled);
   }
 
